server/requirements.txt
flask
Flask_Cors
openai
requests
python-dotenv
psycopg2
jsonschema
firecrawl-py
pinecone-client
beautifulsoup4
tiktoken
